+++ R/oauth-token.r
@@ -218,7 +218,8 @@ oauth2.0_token <- function(endpoint, app, scope = NULL, user_params = NULL,
                            cache = getOption("httr_oauth_cache"),
                            config_init = list(),
                            client_credentials = FALSE,
-                           credentials = NULL
+                           credentials = NULL,
+                           query_authorize_extra = list()
                           ) {
   params <- list(
     scope = scope,
@@ -229,7 +230,8 @@ oauth2.0_token <- function(endpoint, app, scope = NULL, user_params = NULL,
     as_header = as_header,
     use_basic_auth = use_basic_auth,
     config_init = config_init,
-    client_credentials = client_credentials
+    client_credentials = client_credentials,
+    query_authorize_extra = query_authorize_extra
   )
 
   Token2.0$new(
@@ -251,7 +253,9 @@ Token2.0 <- R6::R6Class("Token2.0", inherit = Token, list(
       oob_value = self$params$oob_value,
       use_basic_auth = self$params$use_basic_auth,
       config_init = self$params$config_init,
-      client_credentials = self$params$client_credentials)
+      client_credentials = self$params$client_credentials,
+      query_authorize_extra = self$params$query_authorize_extra
+    )
   },
   can_refresh = function() {
     !is.null(self$credentials$refresh_token)